+++ packages/sampler/stories/moonstone-stories/Marquee.js
@@ -1,4 +1,4 @@
-import {MarqueeText} from '@enact/moonstone/Marquee';
+import Marquee from '@enact/moonstone/Marquee';
 import React from 'react';
 import {storiesOf} from '@storybook/react';
 import {boolean, number, select, text} from '@storybook/addon-knobs';
@@ -13,7 +13,7 @@ storiesOf('Moonstone', module)
 			const disabled = nullify(boolean('disabled', false));
 			return (
 				<section>
-					<MarqueeText
+					<Marquee
 						alignment={nullify(select('alignment', [null, 'left', 'right', 'center']))}
 						disabled={disabled}
 						forceDirection={nullify(select('forceDirection', [null, 'rtl', 'ltr']))}
@@ -26,8 +26,8 @@ storiesOf('Moonstone', module)
 						style={{width: '400px'}}
 					>
 						{text('children', 'The quick brown fox jumped over the lazy dog. The bean bird flies at sundown.')}
-					</MarqueeText>
-					{disabled ? <p style={{fontSize: '70%', fontStyle: 'italic'}}><sup>*</sup>MarqueeText does not visually respond to <code>disabled</code> state.</p> : <p />}
+					</Marquee>
+					{disabled ? <p style={{fontSize: '70%', fontStyle: 'italic'}}><sup>*</sup>Marquee does not visually respond to <code>disabled</code> state.</p> : <p />}
 				</section>
 			);
 		})