+++ packages/ui/CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 - `ui/Transition` property `clipHeight`
 
+### Added
+
+- `ui/Scrollable` support for scrolling by touch
+
 ### Fixed
 
 - `ui/Transition` animation for `clip` for "up", "left", and "right" directions. This includes a DOM addition to the Transition markup.